nm-5/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0
python-dotenv==1.0.1
requests==2.31.0
python-magic==0.4.27
python-magic-bin==0.4.14
django-storages>=1.14.2
gunicorn>=21.2.0
psycopg2-binary>=2.9.9  # For PostgreSQL support
django-filter>=23.5  # For filtering support 
azure-ai-textanalytics==5.3.0
azure.ai.contentsafety== 1.0.0
channels==4.0.0